Addeing CrowdFunding smart contract
This is the Smart Contract used to make the crowd funding on blockchain.
It's often happens in web2 that crowd funding companies(like ngo's) takes the funding from the crowd and use it for their own purposes. Even if the deadline is crossed  they didn't return back the money. So this smart contract helps to solve this problem by using the functionality of smart contract.
Built on:- Ethereum